Be "finito"; Devo implementare la limitazione per tipo di account

+++ NowPuglia-BE/src/main/java/it/Gruppo1/NowPuglia/dto/JsonResponse/JsonDataDto.java
@@ -14,6 +14,6 @@ import java.util.LinkedList;
 @NoArgsConstructor
 @AllArgsConstructor
 public class JsonDataDto {
-    private int misurazioniTotali;
+    private long misurazioniTotali;
     private LinkedList<MisurazioniDto> misurazioni;
 }
